Basic server and moved parse.js
