+++ package.json
@@ -6,15 +6,21 @@
     "server-dev": "npx nodemon --watch server index.js",
     "start": "node index.js",
     "create-db": "psql postgres -f sql/db.sql",
-    "create-tables": "psql qna -f sql/tables.sql"
+    "create-tables": "psql qna -f sql/tables.sql",
+    "etl": "node ETL/answers.js & node ETL/photos.js & node ETL/questions.js"
   },
   "dependencies": {
+    "cors": "^2.8.5",
     "csv": "^6.2.7",
     "csv-parser": "^3.0.0",
     "csv-stream": "^0.2.0",
     "csv-writer": "^1.6.0",
+    "dotenv": "^16.0.3",
+    "express": "^4.18.2",
     "fs": "^0.0.1-security",
+    "morgan": "^1.10.0",
     "ndjson": "^2.0.0",
+    "path": "^0.12.7",
     "pg": "^8.9.0",
     "pg-hstore": "^2.3.4",
     "sequelize": "^6.28.0",
